Rename modules that use customtkinter.  Remove unused variables from modules.
