+++ script bits/getCompanyData_ctk.py
@@ -1,19 +1,15 @@
 import yfinance as yf
-import numDays
-import getTicker
-
-global company_name
-global stockData
+import numDays_ctk
+import getTicker_ctk
 
 def get_data():
-    global dates
     global company_name
     global stockData
-    item = getTicker.ticker
+    item = getTicker_ctk.ticker
     # Pseudo status
     print('Fetching data for', item, '...')
     stockData = yf.download(tickers = item,
-                         start= numDays.dates[2],
-                         end= numDays.dates[3])
+                         start= numDays_ctk.dates[2],
+                         end= numDays_ctk.dates[3])
     cmp = yf.Ticker(item)
     company_name = cmp.info['longName']
